сгруппировал mock-данные в один модуль
