--- src/mock/demo-const.js
@@ -1,5 +1,0 @@
-export const PHOTO_COUNT = 100;
-export const MIN_PRICE = 10;
-export const MAX_PRICE = 100;
-export const POINT_TYPE = 'flight';
-export const POINT_NAME = 'Маршрутная точка (ДЕМО)';
